# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TernakPro API",
        "description": "API untuk aplikasi Ternak yang menyediakan berbagai fitur terkait ternak dan manajemen pengguna.",
        "contact": {
            "email": "dopidev@gmail.com"
        },
        "version": "v1"
    },
    "paths": {
...
        "/api/v1/ternak": {
            "get": {
                "tags": [
                    "Ternak"
                ],
                "summary": "Mengambil semua data ternak",
                "description": "Endpoint ini digunakan untuk mengambil semua data ternak yang tersedia.",
                "operationId": "c9ca4f5964eb347a780d6dd5c0ada6c6",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data ternak",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ternak"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Ternak"
                ],
                "summary": "Menyimpan ternak baru",
                "description": "Endpoint ini digunakan untuk membuat data ternak baru.",
                "operationId": "e157ca2b0dc822bce22eec0eaa8ee66a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "nama_ternak",
                                    "tgl_mulai",
                                    "hewan_id",
                                    "ras_id",
                                    "tujuan_ternak_id",
                                    "usia",
                                    "kondisi_ternak",
                                    "jenis_kelamin",
                                    "berat"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "user123"
                                    },
                                    "nama_ternak": {
                                        "type": "string",
                                        "example": "Sapi Perah"
                                    },
                                    "tgl_mulai": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-08-20"
                                    },
                                    "hewan_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ras_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tujuan_ternak_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "usia": {
                                        "type": "integer",
                                        "example": 24
                                    },
                                    "kondisi_ternak": {
                                        "type": "string",
                                        "example": "Sehat"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string",
                                        "example": "Betina"
                                    },
                                    "berat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500.5
                                    },
                                    "catatan": {
                                        "type": "string",
                                        "example": "Catatan tambahan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ternak berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ternak"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/ternak/{user_id}": {
            "get": {
                "tags": [
                    "Ternak"
                ],
                "summary": "Mengambil semua ternak berdasarkan user_id",
                "description": "Endpoint ini用于 untuk mengambil semua data ternak milik pengguna tertentu berdasarkan user_id.",
                "operationId": "af0dbfad5199a3a374d686e259ba38b5",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID pengguna (uid)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data ternak",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ternak"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/ternak/tujuan/{user_id}": {
            "get": {
                "tags": [
                    "Tujuan Ternak"
                ],
                "summary": "Mengambil semua tujuan ternak berdasarkan user_id",
                "description": "Endpoint ini digunakan untuk mengambil semua data tujuan ternak milik pengguna tertentu berdasarkan user_id.",
                "operationId": "2ebf88bf8ab7f67b2d9c12b0bb1bb47d",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID pengguna (uid)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data tujuan ternak",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TujuanTernak"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tujuan Ternak"
                ],
                "summary": "Mengupdate tujuan ternak",
                "description": "Endpoint ini digunakan untuk mengupdate data tujuan ternak berdasarkan user_id.",
                "operationId": "335f5bab1538c441fc7c674824a5008f",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID pengguna (uid)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "Peternakan Sapi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tujuan ternak berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TujuanTernak"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tujuan ternak tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tujuan Ternak tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tujuan Ternak"
                ],
                "summary": "Menghapus tujuan ternak",
                "description": "Endpoint ini digunakan untuk menghapus data tujuan ternak berdasarkan user_id.",
                "operationId": "80d638a7b9e0b7b24a71701c98a286e4",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID pengguna (uid)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tujuan ternak berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tujuan Ternak berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tujuan ternak tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tujuan Ternak tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TujuanTernak": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "string",
                        "example": "user123"
                    },
                    "nama": {
                        "type": "string",
                        "example": "Peternakan Sapi"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
            "Ternak": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tag_id": {
                        "type": "string",
                        "example": "TAG123"
                    },
                    "user_id": {
                        "type": "string",
                        "example": "user123"
                    },
                    "nama_ternak": {
                        "type": "string",
                        "example": "Sapi Perah"
                    },
                    "tgl_mulai": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-08-20"
                    },
                    "hewan_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ras_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tujuan_ternak_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "usia": {
                        "type": "integer",
                        "example": 24
                    },
                    "kondisi_ternak": {
                        "type": "string",
                        "example": "Sehat"
                    },
                    "jenis_kelamin": {
                        "type": "string",
                        "example": "Betina"
                    },
                    "berat": {
                        "type": "number",
                        "format": "float",
                        "example": 500.5
                    },
                    "catatan": {
                        "type": "string",
                        "example": "Catatan tambahan"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "hewan": {
                        "$ref": "#/components/schemas/Hewan"
                    },
                    "ras": {
                        "$ref": "#/components/schemas/Ras"
                    },
                    "tujuan_ternak": {
                        "$ref": "#/components/schemas/TujuanTernak"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Masukkan token dalam format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}